added more inline typing
