+++ src/utils/getRepoByLanguage.ts
@@ -6,6 +6,8 @@ import { Octokit } from "./octokitTypes";
 
 import { searchParameters, searchResponse } from "./octokitTypes";
 
+import { response, usersWriteAdminReposArray } from "../../types/common";
+
 export const fetchReposByLanguage = async (
   octokit: Octokit
 ): Promise<response> => {
