+++ src/utils/writeToFile.ts
@@ -1,5 +1,7 @@
 import { error, inform } from "./globals";
 
+import { usersWriteAdminReposArray, response } from "../../types/common"
+
 import { promisify } from "util";
 
 import fs from "fs";
