--- types/common/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "common",
-  "version": "1.0.0",
-  "typings": "main.d.ts"
-}
